tipos de dados das coordenadas de quadros e tamanho de quadros foram alterados para double

+++ server/src/PushIt/DTOs_Contracts/UpdateQuadroRequest.cs
@@ -1,9 +1,9 @@
 public record UpdateQuadroRequest
 (
-     int x,
-     int y,
-     int width,
-     int height,
+     double x,
+     double y,
+     double width,
+     double height,
      string text,
      string colour
 );
